requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
reportlab>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pypdf>=3.17.0
posthog>=3.1.0
